registry: preserve annotation object identity across requests
Previously, updating references to annotation objects was a laborious task that
needed to be done manually. With a registry to take care of this task,
annotation objects themselves can be preserved across remote requests, so that
anyone who has a reference to an annotation, has that reference for all time.

Also, deduplicate management of the annotations in the registry.