core: Pull annotation registry tasks out of Annotator
This commit starts the work of extricating the job of tracking annotation
storage state from Annotator core. The Store plugin is dramatically simplified,
leaving all the event publishing to the Registry, and large chunks of Annotator
core itself can be removed.

This breaks large chunks of the Store and Annotator specs -- I'll fix that in a
future commit =).